[audio hal] delete unused variant in struct pcm_config
The variant flag is passed to hdmi dirver to tell hdmi
which type/format audio stream will be written. This
lead to modify codes of tinyalsa and it is not a good
way. The driver of hdmi in kernel 4.19 not support
this way(pass a flag to driver), so delete it to
compile with the tinyalsa.

Change-Id: I60640772b0d88605fff3599769c2be9f6aaddda5
Signed-off-by: He Hua <hh@rock-chips.com>